chore: organize loose DataAccess layer classes
